+++ Source/WebCore/rendering/RenderRubyRun.h
@@ -63,7 +63,7 @@ public:
     virtual RenderBlock* firstLineBlock() const override;
     virtual void updateFirstLetter() override;
 
-    void getOverhang(bool firstLine, RenderObject* startRenderer, RenderObject* endRenderer, int& startOverhang, int& endOverhang) const;
+    void getOverhang(bool firstLine, RenderObject* startRenderer, RenderObject* endRenderer, float& startOverhang, float& endOverhang) const;
 
     static RenderRubyRun* staticCreateRubyRun(const RenderObject* parentRuby);
